New line at EOF + more small compile warning fixes.

+++ cocos2d/CCLabelAtlas.m
@@ -70,7 +70,7 @@
 
 -(void) updateAtlasValues
 {
-	int n = [string_ length];
+	NSUInteger n = [string_ length];
 	
 	ccV3F_C4B_T2F_Quad quad;
